EbThreads: Fix pthread error handling
When creating a thread fails, no new thread is created and the contents
of the thread handle are undefined. Using it to call pthread_cancel
is undefined behavior therefore. There is no point cancelling if no
thread was actually created.

Additionally in case of EPERM error, there is no need to free the
memory allocated for the thread handle, just to allocate it again
right after.

When re-trying to create the thread without the scheduling attributes,
error handling was missing which could lead to returning an undefined
thread handle, so this has to be handled correctly too.