EB_API: Use a def file instead of __declspec(dllexport)
Both results in the same output, however, with the use of `__declspec(dllexport)` without
having proper guards when compiling the same headers in the executable, CMake and by extension
Visual Studio will generate a .lib file for SvtAv1EncApp and SvtAv1DecApp, which is not wanted.

The alternative is to trace the headers and add extra guards to prevent
`__declspec(dllexport)` from appearing when compiling the executables

The def file is generated by running `gendef - SvtAv1Enc.dll > SvtAv1Enc.def`
using gendef from mingw-w64 gotten through msys2.
To add to the def file, just add the function name to the def file as long as the function
name is not mangled like in C++.

https://github.com/Alexpux/mingw-w64/tree/master/mingw-w64-tools/gendef
https://docs.microsoft.com/en-us/cpp/build/reference/module-definition-dot-def-files?view=vs-2019
https://docs.microsoft.com/en-us/cpp/build/exporting-from-a-dll-using-def-files?view=vs-2019
https://docs.microsoft.com/en-us/cpp/build/reference/exports?view=vs-2019
https://docs.microsoft.com/en-us/cpp/cpp/dllexport-dllimport?view=vs-2019
https://docs.microsoft.com/en-us/cpp/build/reference/export-exports-a-function?view=vs-2019

Signed-off-by: Christopher Degawa <christopher.degawa@intel.com>